// Haly.WebApp/ClientApp/generated/haly/models/DuplicatesStrategy.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Haly API
 * An ASP.NET Core Web API that adds quality of life improvements to Spotify and helps with music exploration
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const DuplicatesStrategy = {
    FailWhenAnyDuplicate: 'FailWhenAnyDuplicate',
    AddNewOnes: 'AddNewOnes',
    AddAll: 'AddAll'
} as const;
export type DuplicatesStrategy = typeof DuplicatesStrategy[keyof typeof DuplicatesStrategy];


export function DuplicatesStrategyFromJSON(json: any): DuplicatesStrategy {
    return DuplicatesStrategyFromJSONTyped(json, false);
}

export function DuplicatesStrategyFromJSONTyped(json: any, ignoreDiscriminator: boolean): DuplicatesStrategy {
    return json as DuplicatesStrategy;
}

export function DuplicatesStrategyToJSON(value?: DuplicatesStrategy | null): any {
    return value as any;
}